- added validation on regsiterPage.js - added alert on registerPage.js - get expoPushToken when user register
