+++ package.json
@@ -13,6 +13,8 @@
     "@react-navigation/stack": "^5.9.3",
     "axios": "^0.20.0",
     "expo": "~39.0.2",
+    "expo-notifications": "^0.7.2",
+    "expo-permissions": "^9.3.0",
     "expo-status-bar": "~1.0.2",
     "react": "16.13.1",
     "react-dom": "16.13.1",
