+++ src/components/Loading.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Loading() {
+    return (
+        <div>
+            Loading .....
+        </div>
+    )
+}
